# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FVOCI API",
    "description": "Rust slice HTTP contract for authentication and workspace operations.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "logout",
        "responses": {
          "204": {
            "description": "Logged out"
          }
        },
        "security": [
          {
            "fvoci_session": []
          }
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "operationId": "me_get",
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionUserOutput"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "fvoci_session": []
          }
        ]
      },
      "patch": {
        "tags": [
          "auth"
        ],
        "operationId": "me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchMeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionUserOutput"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "fvoci_session": []
          }
        ]
      }
    },
    "/api/v1/me/personal-workspace": {
      "post": {
        "tags": [
          "workspaces"
        ],
        "operationId": "personal_workspace",
        "responses": {
          "200": {
            "description": "Personal workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceMetaResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "fvoci_session": []
          }
        ]
      }
    },
    "/api/v1/me/workspaces": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "operationId": "list_my_workspaces",
        "responses": {
          "200": {
            "description": "Accessible workspaces",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "fvoci_session": []
          }
        ]
      }
    },
    "/api/v1/setup": {
      "get": {
        "tags": [
          "setup"
        ],
        "operationId": "setup_status",
        "responses": {
          "200": {
            "description": "Setup status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupStatusResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "setup"
        ],
        "operationId": "setup_run",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          },
          "404": {
            "description": "Already completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          },
          "409": {
            "description": "Slug taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces": {
      "post": {
        "tags": [
          "workspaces"
        ],
        "operationId": "create_workspace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceMetaResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          },
          "409": {
            "description": "Slug taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "fvoci_session": []
          }
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "operationId": "get_workspace",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceMetaResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found or forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "fvoci_session": []
          }
        ]
      },
      "patch": {
        "tags": [
          "workspaces"
        ],
        "operationId": "patch_workspace",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchWorkspaceBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceMetaResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found or forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "fvoci_session": []
          }
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/members/{user_id}": {
      "delete": {
        "tags": [
          "workspaces"
        ],
        "operationId": "remove_member",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "description": "Member user id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Removed member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found or forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "fvoci_session": []
          }
        ]
      },
      "patch": {
        "tags": [
          "workspaces"
        ],
        "operationId": "patch_member",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "description": "Member user id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberRoleBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found or forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "fvoci_session": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "BrandingOutput": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "CreateWorkspaceBody": {
        "type": "object",
        "required": [
          "name",
          "slug"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "LoginBody": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "LoginResponse": {
        "type": "object",
        "required": [
          "userId"
        ],
        "properties": {
          "userId": {
            "type": "string"
          }
        }
      },
      "MemberResponse": {
        "type": "object",
        "required": [
          "userId",
          "email",
          "givenName",
          "role"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "familyName": {
            "type": [
              "string",
              "null"
            ]
          },
          "givenName": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          }
        }
      },
      "MemberRoleBody": {
        "type": "object",
        "required": [
          "role"
        ],
        "properties": {
          "role": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "OkResponse": {
        "type": "object",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      },
      "PatchMeBody": {
        "type": "object",
        "description": "PATCH /api/v1/auth/me body. `familyName` omitted preserves the value; null clears it.",
        "required": [
          "givenName"
        ],
        "properties": {
          "familyName": {
            "type": [
              "string",
              "null"
            ]
          },
          "givenName": {
            "type": "string"
          },
          "locale": {
            "type": [
              "string",
              "null"
            ]
          },
          "textScale": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "timezone": {
            "type": [
              "string",
              "null"
            ]
          },
          "weekStartsOn": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "PatchWorkspaceBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "additionalProperties": false
      },
      "ProblemResponse": {
        "type": "object",
        "required": [
          "title",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "params": {},
          "source": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": "string"
          }
        }
      },
      "SessionUserOutput": {
        "type": "object",
        "required": [
          "userId",
          "email",
          "givenName",
          "textScale",
          "sessionId",
          "hasPassword",
          "isInstanceAdmin",
          "locale",
          "timezone",
          "weekStartsOn"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "emailVerifiedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "familyName": {
            "type": [
              "string",
              "null"
            ]
          },
          "givenName": {
            "type": "string"
          },
          "hasPassword": {
            "type": "boolean"
          },
          "isInstanceAdmin": {
            "type": "boolean"
          },
          "locale": {
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          },
          "textScale": {
            "type": "integer",
            "format": "int32"
          },
          "timezone": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "weekStartsOn": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "SetupBody": {
        "type": "object",
        "required": [
          "email",
          "password",
          "givenName",
          "workspaceSlug",
          "workspaceName"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "familyName": {
            "type": [
              "string",
              "null"
            ]
          },
          "givenName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "workspaceName": {
            "type": "string"
          },
          "workspaceSlug": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "SetupResponse": {
        "type": "object",
        "required": [
          "userId",
          "workspaceId"
        ],
        "properties": {
          "userId": {
            "type": "string"
          },
          "workspaceId": {
            "type": "string"
          }
        }
      },
      "SetupStatusResponse": {
        "type": "object",
        "required": [
          "needed",
          "branding"
        ],
        "properties": {
          "branding": {
            "$ref": "#/components/schemas/BrandingOutput"
          },
          "needed": {
            "type": "boolean"
          }
        }
      },
      "WorkspaceListItemResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "role",
          "kind",
          "documentCount",
          "assignedCount"
        ],
        "properties": {
          "assignedCount": {
            "type": "integer",
            "format": "int32"
          },
          "documentCount": {
            "type": "integer",
            "format": "int32"
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        }
      },
      "WorkspaceListResponse": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkspaceListItemResponse"
            }
          }
        }
      },
      "WorkspaceMetaResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "slug"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "fvoci_session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "fvoci_session"
      }
    }
  },
  "tags": [
    {
      "name": "setup",
      "description": "Instance setup"
    },
    {
      "name": "auth",
      "description": "Authentication and profile"
    },
    {
      "name": "workspaces",
      "description": "Workspace membership and metadata"
    }
  ]
}